+++ src/Inference/MC/PMMH.hs
@@ -65,7 +65,7 @@ handleModel n prog τθ  = do
 
 {- | An acceptance mechanism for PMMH.
 -}
-handleAccept :: Member Sampler fs => Handler (Accept LogP) fs a a
+handleAccept :: Member Sampler fs => Handler (Proposal LogP) fs a a
 handleAccept (Val x)   = pure x
 handleAccept (Op op k) = case discharge op of
   Right (Propose τθ)